# chittyos-openai-gpt/chatgpt_server.py
#!/usr/bin/env python3
"""
ChittyOS ChatGPT Integration Server
FastAPI server that implements OpenAPI spec for Custom GPT Actions
"""

import os
import json
import logging
from datetime import datetime, timezone
from typing import Dict, List, Optional, Any
from enum import Enum

import httpx
import uvicorn
from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends, Security
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel, Field

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger("chittyos-gpt")

# FastAPI app
app = FastAPI(
    title="ChittyOS Infrastructure API for ChatGPT",
    description="AI-driven legal tech infrastructure management",
    version="1.0.0",
    servers=[
        {"url": "https://api.chitty.cc/v1", "description": "Production"},
        {"url": "http://localhost:8000", "description": "Development"}
    ]
)

# CORS configuration for ChatGPT
app.add_middleware(
    CORSMiddleware,
    allow_origins=["https://chat.openai.com"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Security
security = HTTPBearer()

def verify_token(credentials: HTTPAuthorizationCredentials = Security(security)):
    """Verify API token"""
    token = credentials.credentials
    expected_token = os.getenv("CHITTY_API_KEY")
    if not expected_token or token != expected_token:
        raise HTTPException(status_code=403, detail="Invalid authentication")
    return token

# ============= DATA MODELS =============

class Environment(str, Enum):
    development = "development"
    staging = "staging"
    production = "production"

class CaseType(str, Enum):
    litigation = "litigation"
    contract = "contract"
    merger = "merger"
    compliance = "compliance"
    intellectual_property = "intellectual_property"

class WorkerDeployRequest(BaseModel):
    script_name: str
    script_content: str
    case_id: Optional[str] = None
    environment: Environment = Environment.production
    compliance_requirements: Optional[List[str]] = None

class R2BucketRequest(BaseModel):
    bucket_name: str
    case_id: Optional[str] = None
    encryption: bool = True
    lifecycle_rules: Optional[Dict] = None

class D1DatabaseRequest(BaseModel):
    database_name: str
    case_id: Optional[str] = None
    schema: Optional[str] = None

class LegalCaseRequest(BaseModel):
    client_name: str
    case_type: CaseType
    jurisdiction: Optional[str] = "US"
    priority: str = "normal"
    auto_provision: bool = True

class TransactionRequest(BaseModel):
    amount: float
    from_account: str = Field(alias="from")
    to_account: str = Field(alias="to")
    currency: str = "USD"
    type: str = "transfer"
    case_id: Optional[str] = None

class PropertyListingRequest(BaseModel):
    address: str
    price: float
    type: str
    details: Optional[Dict] = None

class AssetTokenizationRequest(BaseModel):
    asset_name: str
    asset_type: str
    value: float
    token_standard: str = "ERC721"
    total_supply: int = 1

# ============= CHITTYOS INTEGRATION =============

class ChittyOSConnector:
    """Connector to ChittyOS services"""
    
    def __init__(self):
        self.cloudflare_api = os.getenv("CLOUDFLARE_API_TOKEN")
        self.cloudflare_account = os.getenv("CLOUDFLARE_ACCOUNT_ID")
        self.services = {
            "cases": os.getenv("CHITTY_CASE_DB_URL", "https://cases.chitty.cc/api"),
            "chain": os.getenv("CHITTY_CHAIN_RPC", "https://chain.chitty.cc/rpc"),
            "verify": os.getenv("CHITTY_VERIFY_ENDPOINT", "https://verify.chitty.cc/api"),
            "trust": os.getenv("CHITTY_TRUST_API", "https://trust.chitty.cc/api"),
            "gateway": os.getenv("CHITTY_GATEWAY_URL", "https://gateway.chitty.cc")
        }
    
    async def deploy_worker(self, request: WorkerDeployRequest) -> Dict:
        """Deploy Worker to Cloudflare with ChittyOS tracking"""
        # Simulate deployment for demo
        deployment_id = f"wkr-{datetime.now(timezone.utc).strftime('%Y%m%d%H%M%S')}"
        
        # In production, this would call Cloudflare API
        result = {
            "deployment_id": deployment_id,
            "script_name": request.script_name,
            "status": "deployed",
            "case_link": {"case_id": request.case_id, "linked": True} if request.case_id else None,
            "blockchain": {"tx_hash": f"0x{'0' * 64}", "anchored": True},
            "compliance": {"compliant": True, "checks_passed": request.compliance_requirements or []},
            "url": f"https://{request.script_name}.workers.dev"
        }
        
        logger.info(f"Deployed worker: {deployment_id}")
        return result
    
    async def create_legal_case(self, request: LegalCaseRequest) -> Dict:
        """Create legal case with infrastructure provisioning"""
        case_id = f"CASE-{datetime.now(timezone.utc).strftime('%Y%m%d')}-{request.client_name[:3].upper()}"
        chitty_id = f"CHITTY-{case_id}-{datetime.now(timezone.utc).timestamp()}"
        
        result = {
            "case_id": case_id,
            "chitty_id": chitty_id,
            "status": "created",
            "infrastructure": {}
        }
        
        if request.auto_provision:
            # Auto-provision infrastructure
            result["infrastructure"] = {
                "worker": f"{case_id.lower()}-worker",
                "database": f"{case_id.lower()}-db",
                "storage": f"{case_id.lower()}-docs"
            }
        
        logger.info(f"Created case: {case_id}")
        return result
    
    async def anchor_to_blockchain(self, data: Dict, type: str) -> Dict:
        """Anchor data to ChittyChain"""
        tx_hash = f"0x{''.join([str(ord(c) % 16) for c in str(data)[:64]]).ljust(64, '0')}"
        
        result = {
            "transaction_hash": tx_hash,
            "block_number": 12345678,
            "timestamp": datetime.now(timezone.utc).isoformat(),
            "network": "chittychain-mainnet"
        }
        
        logger.info(f"Anchored to blockchain: {tx_hash}")
        return result

connector = ChittyOSConnector()

# ============= API ENDPOINTS =============

@app.post("/cloudflare/workers/deploy")
async def deploy_worker(
    request: WorkerDeployRequest,
    token: str = Depends(verify_token)
):
    """Deploy Cloudflare Worker with ChittyOS tracking"""
    try:
        result = await connector.deploy_worker(request)
        return result
    except Exception as e:
        logger.error(f"Deployment failed: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/cloudflare/r2/buckets")
async def create_r2_bucket(
    request: R2BucketRequest,
    token: str = Depends(verify_token)
):
    """Create R2 storage bucket"""
    bucket_id = f"r2-{datetime.now(timezone.utc).strftime('%Y%m%d%H%M%S')}"
    
    return {
        "bucket_id": bucket_id,
        "bucket_name": request.bucket_name,
        "status": "created",
        "case_id": request.case_id,
        "endpoint": f"https://{request.bucket_name}.r2.cloudflarestorage.com"
    }

@app.post("/cloudflare/d1/databases")
async def create_d1_database(
    request: D1DatabaseRequest,
    token: str = Depends(verify_token)
):
    """Create D1 database"""
    database_id = f"d1-{datetime.now(timezone.utc).strftime('%Y%m%d%H%M%S')}"
    
    return {
        "database_id": database_id,
        "database_name": request.database_name,
        "status": "created",
        "case_id": request.case_id
    }

@app.post("/legal/cases")
async def create_legal_case(
    request: LegalCaseRequest,
    token: str = Depends(verify_token)
):
    """Create legal case with full infrastructure provisioning"""
    try:
        result = await connector.create_legal_case(request)
        return result
    except Exception as e:
        logger.error(f"Case creation failed: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/legal/documents/analyze")
async def analyze_document(
    document: UploadFile = File(...),
    case_id: str = Form(...),
    analysis_type: str = Form("contract_review"),
    token: str = Depends(verify_token)
):
    """AI-powered analysis of legal documents"""
    document_id = f"doc-{datetime.now(timezone.utc).strftime('%Y%m%d%H%M%S')}"
    
    # Simulated analysis
    return {
        "document_id": document_id,
        "case_id": case_id,
        "risk_score": 75.5,
        "compliance_status": "compliant",
        "key_findings": [
            "Liability clause needs review",
            "Termination conditions are standard",
            "IP assignment is comprehensive"
        ],
        "recommendations": [
            "Add arbitration clause",
            "Specify jurisdiction explicitly",
            "Include force majeure provisions"
        ]
    }

@app.post("/finance/transactions")
async def create_transaction(
    request: TransactionRequest,
    token: str = Depends(verify_token)
):
    """Create financial transaction"""
    transaction_id = f"TXN-{datetime.now(timezone.utc).strftime('%Y%m%d%H%M%S')}"
    chitty_id = f"CHITTY-TXN-{transaction_id}"
    
    return {
        "transaction_id": transaction_id,
        "chitty_id": chitty_id,
        "status": "pending",
        "amount": request.amount,
        "blockchain_tx": f"0x{'0' * 64}"
    }

@app.post("/property/listings")
async def create_property_listing(
    request: PropertyListingRequest,
    token: str = Depends(verify_token)
):
    """Create property listing"""
    listing_id = f"PROP-{datetime.now(timezone.utc).strftime('%Y%m%d%H%M%S')}"
    chitty_id = f"CHITTY-PROP-{listing_id}"
    
    # Simulated valuation
    estimated_value = request.price * (1 + (0.1 * (0.5 - 0.5)))  # ±10% adjustment
    
    return {
        "listing_id": listing_id,
        "chitty_id": chitty_id,
        "status": "active",
        "valuation": {
            "estimated_value": estimated_value,
            "confidence": 0.85,
            "comparables": 3
        }
    }

@app.post("/assets/tokenize")
async def tokenize_asset(
    request: AssetTokenizationRequest,
    token: str = Depends(verify_token)
):
    """Tokenize asset on blockchain"""
    token_id = f"TKN-{datetime.now(timezone.utc).strftime('%Y%m%d%H%M%S')}"
    
    return {
        "token_id": token_id,
        "contract_address": f"0x{'a' * 40}",
        "transaction_hash": f"0x{'b' * 64}",
        "total_supply": request.total_supply
    }

@app.post("/chain/anchor")
async def anchor_to_blockchain(
    data: Dict = None,
    type: str = None,
    case_id: Optional[str] = None,
    token: str = Depends(verify_token)
):
    """Anchor data to blockchain"""
    try:
        result = await connector.anchor_to_blockchain(data, type)
        return result
    except Exception as e:
        logger.error(f"Blockchain anchoring failed: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/compliance/verify")
async def verify_compliance(
    resource: Dict = None,
    requirements: List[str] = None,
    token: str = Depends(verify_token)
):
    """Verify compliance requirements"""
    
    # Simulated compliance check
    compliant = True
    issues = []
    
    if "GDPR" in requirements and "data_retention" not in resource:
        issues.append("Missing GDPR data retention policy")
        compliant = False
    
    return {
        "compliant": compliant,
        "score": 85.0 if compliant else 60.0,
        "issues": issues,
        "recommendations": [
            "Implement data retention policy" if issues else "All checks passed"
        ]
    }

@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return {
        "status": "healthy",
        "service": "ChittyOS ChatGPT Integration",
        "timestamp": datetime.now(timezone.utc).isoformat()
    }

@app.get("/openapi.json")
async def get_openapi():
    """Return OpenAPI spec for ChatGPT"""
    return app.openapi()

if __name__ == "__main__":
    port = int(os.getenv("PORT", 8000))
    uvicorn.run(
        "chatgpt_server:app",
        host="0.0.0.0",
        port=port,
        reload=True,
        log_level="info"
    )